requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
ics>=0.7.2
python-dateutil>=2.9.0
google-api-python-client>=2.118.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
